Split transforms.cpp into several files
