+++ deps/src/Makefile
@@ -7,11 +7,17 @@ LIBRARIES = -L$(HEALPIX)/lib -L$(HEALPIX)/src/cxx/gcc_with_fpic/lib -lhealpix_cx
 
 all: libhealpixwrapper.so
 
+map.o: map.cpp
+	$(CXX) $(CXXFLAGS) map.cpp $(INCLUDES)
+
+alm.o: alm.cpp
+	$(CXX) $(CXXFLAGS) alm.cpp $(INCLUDES)
+
 transforms.o: transforms.cpp
 	$(CXX) $(CXXFLAGS) transforms.cpp $(INCLUDES)
 
-libhealpixwrapper.so: transforms.o
-	$(CXX) -shared -o libhealpixwrapper.so transforms.o $(LIBRARIES)
+libhealpixwrapper.so: map.o alm.o transforms.o
+	$(CXX) -shared -o libhealpixwrapper.so map.o alm.o transforms.o $(LIBRARIES)
 
 clean:
 	-rm *.o
